Lots of bugfixes
But still a segmentation fault D:

+++ Plane.cpp
@@ -2,7 +2,7 @@
 #include "Ray.h"
 
 Collision Plane::collide(const Ray& _ray) const{
-    int t,temp2,temp3;
+    float t,temp2,temp3;
     glm::vec3 temp1,collisionPoint,collisionNormal;
     const Material* planeMaterial;
     //vec3s _o _d _n _a
